+++ lib/auth.tsx
@@ -2,7 +2,7 @@
 
 import { createContext, useContext, useEffect, useState } from 'react'
 import { User, Session, AuthChangeEvent } from '@supabase/supabase-js'
-import { supabase } from './supabase'
+import { supabase } from '@/lib/db/client'
 
 interface AuthContextType {
   user: User | null
