+++ lib/db/client.ts
@@ -1,0 +1,27 @@
+'use client'
+
+import { createClient } from '@supabase/supabase-js'
+
+const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL as string | undefined
+const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY as string | undefined
+
+function createStubClient() {
+  const msg = '[Supabase] NEXT_PUBLIC_SUPABASE_URL or NEXT_PUBLIC_SUPABASE_ANON_KEY is missing. Configure your environment variables (e.g., in .env.local or Vercel Project Settings).'
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  return new Proxy({}, {
+    get() {
+      throw new Error(msg)
+    },
+    apply() {
+      throw new Error(msg)
+    },
+  }) as any
+}
+
+if (!supabaseUrl || !supabaseAnonKey) {
+  console.warn('[Supabase] Missing NEXT_PUBLIC_SUPABASE_URL or NEXT_PUBLIC_SUPABASE_ANON_KEY. Set them in .env.local or Vercel Project Settings.')
+}
+
+export const supabase = (supabaseUrl && supabaseAnonKey)
+  ? createClient(supabaseUrl, supabaseAnonKey)
+  : createStubClient()
